+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and `Removed`.
 ### Added
 
 - Ability to select compression level for when using Zstd backups.
+- Ability to backup Fonts folder.
 
 ## [1.2.2] - 2021-07-06